# OPENAPI/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paperless-AI API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Paperless-AI application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Clusterzx",
      "url": "https://github.com/Clusterzx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT-based authentication for web app users. The token is obtained by authenticating via the login endpoint.\n\n### How to authenticate:\n1. Send a POST request to `/login` with your username and password\n2. The server will respond with a JWT token (also set as a cookie in browsers)\n3. Include this token in the `Authorization` header as `Bearer {token}`\n\nExample:\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nJWT tokens are valid for 24 hours after issuance.\n"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key authentication for programmatic access. The API key can be generated or regenerated using the /api/key-regenerate endpoint.\n\n### How to authenticate:\n1. Access the API key from your application settings\n2. Include the API key in the `x-api-key` HTTP header for all requests\n\nExample:\n```\nx-api-key: 7c1f3f5e2b0a9d8c6e4b2a1d3f5e8c9b2a1d3f5e\n```\n\nAPI keys do not expire unless regenerated.\n"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Error resetting documents"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/login": {
      "get": {
        "summary": "Render login page or redirect to setup if no users exist",
        "description": "Serves the login page for user authentication to the Paperless-AI application.\nIf no users exist in the database, the endpoint automatically redirects to the setup page\nto complete the initial application configuration.\n\nThis endpoint handles both new user sessions and returning users whose\nsessions have expired.\n",
        "tags": [
          "Authentication",
          "Navigation"
        ],
        "responses": {
          "200": {
            "description": "Login page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "HTML content of the login page"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to setup page if no users exist, or to dashboard if already authenticated",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/setup"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Authenticate user with username and password",
        "description": "Authenticates a user using their username and password credentials.\nIf authentication is successful, a JWT token is generated and stored in a secure HTTP-only\ncookie for subsequent requests.\n\nFailed login attempts are logged for security purposes, and multiple failures\nmay result in temporary account lockout depending on configuration.\n",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "User's login name",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password",
                    "example": "securepassword"
                  },
                  "rememberMe": {
                    "type": "boolean",
                    "description": "Whether to extend the session lifetime",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "redirect": {
                      "type": "string",
                      "description": "URL to redirect to after successful login",
                      "example": "/dashboard"
                    }
                  }
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "HTTP-only cookie containing JWT token"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid username or password"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "get": {
        "summary": "Log out user and clear JWT cookie",
        "description": "Terminates the current user session by invalidating and clearing the JWT authentication\ncookie. After logging out, the user is redirected to the login page.\n\nThis endpoint also clears any session-related data stored on the server side\nfor the current user.\n",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "302": {
            "description": "Logout successful, redirected to login page",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/login"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "description": "HTTP-only cookie with cleared JWT token and immediate expiration"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "summary": "Application settings page",
        "description": "Renders the application settings page where users can modify configuration\nafter initial setup.\n\nThis page allows administrators to update connections to Paperless-ngx,\nAI provider settings, processing parameters, feature toggles, and custom fields.\nThe interface provides validation for connection settings and displays the current\nconfiguration values.\n\nChanges made on this page require application restart to take full effect.\n",
        "tags": [
          "Navigation",
          "Setup",
          "System"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "HTML content of the application settings page"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/login"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update application settings",
        "description": "Updates the configuration settings of the Paperless-AI application after initial setup.\nThis endpoint allows administrators to modify connections to Paperless-ngx,\nAI provider settings, processing parameters, and feature toggles.\n\nChanges made through this endpoint are applied immediately and affect all future\ndocument processing operations.\n",
        "tags": [
          "System",
          "Setup"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paperlessUrl": {
                    "type": "string",
                    "description": "URL of the Paperless-ngx instance",
                    "example": "https://paperless.example.com"
                  },
                  "paperlessToken": {
                    "type": "string",
                    "description": "API token for Paperless-ngx access",
                    "example": "abc123def456"
                  },
                  "paperlessUsername": {
                    "type": "string",
                    "description": "Username for Paperless-ngx (alternative to token authentication)",
                    "example": "admin"
                  },
                  "aiProvider": {
                    "type": "string",
                    "description": "Selected AI provider for document analysis",
                    "enum": [
                      "openai",
                      "ollama",
                      "custom",
                      "azure"
                    ],
                    "example": "openai"
                  },
                  "openaiKey": {
                    "type": "string",
                    "description": "API key for OpenAI (required when aiProvider is 'openai')",
                    "example": "sk-abc123def456"
                  },
                  "openaiModel": {
                    "type": "string",
                    "description": "OpenAI model to use for analysis",
                    "example": "gpt-4"
                  },
                  "ollamaUrl": {
                    "type": "string",
                    "description": "URL for Ollama API (required when aiProvider is 'ollama')",
                    "example": "http://localhost:11434"
                  },
                  "ollamaModel": {
                    "type": "string",
                    "description": "Ollama model to use for analysis",
                    "example": "llama2"
                  },
                  "customApiKey": {
                    "type": "string",
                    "description": "API key for custom LLM provider",
                    "example": "api-key-123"
                  },
                  "customBaseUrl": {
                    "type": "string",
                    "description": "Base URL for custom LLM provider",
                    "example": "https://api.customllm.com"
                  },
                  "customModel": {
                    "type": "string",
                    "description": "Model name for custom LLM provider",
                    "example": "custom-model"
                  },
                  "scanInterval": {
                    "type": "number",
                    "description": "Interval in minutes for scanning new documents",
                    "example": 15
                  },
                  "systemPrompt": {
                    "type": "string",
                    "description": "Custom system prompt for document analysis",
                    "example": "Extract key information from the following document..."
                  },
                  "showTags": {
                    "type": "boolean",
                    "description": "Whether to show tags in the UI",
                    "example": true
                  },
                  "tokenLimit": {
                    "type": "integer",
                    "description": "The maximum number of tokens th AI can handle",
                    "example": 128000
                  },
                  "responseTokens": {
                    "type": "integer",
                    "description": "The approx. amount of tokens required for the response",
                    "example": 1000
                  },
                  "tags": {
                    "type": "string",
                    "description": "Comma-separated list of tags to use for filtering",
                    "example": "Invoice,Receipt,Contract"
                  },
                  "aiProcessedTag": {
                    "type": "boolean",
                    "description": "Whether to add a tag for AI-processed documents",
                    "example": true
                  },
                  "aiTagName": {
                    "type": "string",
                    "description": "Tag name to use for AI-processed documents",
                    "example": "AI-Processed"
                  },
                  "usePromptTags": {
                    "type": "boolean",
                    "description": "Whether to use tags in prompts",
                    "example": true
                  },
                  "promptTags": {
                    "type": "string",
                    "description": "Comma-separated list of tags to use in prompts",
                    "example": "Invoice,Receipt"
                  },
                  "useExistingData": {
                    "type": "boolean",
                    "description": "Whether to use existing data from a previous setup",
                    "example": false
                  },
                  "activateTagging": {
                    "type": "boolean",
                    "description": "Enable AI-based tag suggestions",
                    "example": true
                  },
                  "activateCorrespondents": {
                    "type": "boolean",
                    "description": "Enable AI-based correspondent suggestions",
                    "example": true
                  },
                  "activateDocumentType": {
                    "type": "boolean",
                    "description": "Enable AI-based document type suggestions",
                    "example": true
                  },
                  "activateTitle": {
                    "type": "boolean",
                    "description": "Enable AI-based title suggestions",
                    "example": true
                  },
                  "activateCustomFields": {
                    "type": "boolean",
                    "description": "Enable AI-based custom field extraction",
                    "example": false
                  },
                  "customFields": {
                    "type": "string",
                    "description": "JSON string defining custom fields to extract",
                    "example": "{\"invoice_number\":{\"type\":\"string\"},\"total_amount\":{\"type\":\"number\"}}"
                  },
                  "disableAutomaticProcessing": {
                    "type": "boolean",
                    "description": "Disable automatic document processing",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "success"
                      ],
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Settings updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ],
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid settings: AI provider required when automatic processing is enabled"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while updating settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ],
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to update settings: Database error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}